% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-dependent.R
\name{asian_call_payoff}
\alias{asian_call_payoff}
\title{Arithmetic-average Asian call payoff}
\usage{
asian_call_payoff(strike, include_spot = FALSE)
}
\arguments{
\item{strike}{Strike price.}

\item{include_spot}{Whether the initial spot column participates in the average.}
}
\value{
A payoff function accepting a path matrix.
}
\description{
Arithmetic-average Asian call payoff
}
