% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-scholes.R
\name{black_scholes_price}
\alias{black_scholes_price}
\title{Black-Scholes price for a European option}
\usage{
black_scholes_price(
  spot,
  strike,
  maturity,
  rate,
  volatility,
  dividend_yield = 0,
  type = c("call", "put")
)
}
\arguments{
\item{spot}{Current underlying price.}

\item{strike}{Strike price.}

\item{maturity}{Time to maturity in years.}

\item{rate}{Continuously compounded risk-free rate.}

\item{volatility}{Annualized volatility.}

\item{dividend_yield}{Continuous dividend yield.}

\item{type}{Either \code{"call"} or \code{"put"}.}
}
\value{
A numeric option price.
}
\description{
Computes the analytical Black-Scholes-Merton price of a European call or put
with a continuous dividend yield.
}
