% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-dependent.R
\name{down_and_out_call_payoff}
\alias{down_and_out_call_payoff}
\title{Down-and-out European call payoff}
\usage{
down_and_out_call_payoff(strike, barrier)
}
\arguments{
\item{strike}{Strike price.}

\item{barrier}{Lower barrier level.}
}
\value{
A payoff function accepting a path matrix.
}
\description{
Down-and-out European call payoff
}
