% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-scholes.R
\name{payoff_helpers}
\alias{payoff_helpers}
\alias{call_payoff}
\alias{put_payoff}
\alias{digital_call_payoff}
\title{Standard terminal payoff functions}
\usage{
call_payoff(strike)

put_payoff(strike)

digital_call_payoff(strike, cash = 1)
}
\arguments{
\item{strike}{Strike price.}

\item{cash}{Cash amount paid by the digital call when in the money.}
}
\value{
A vectorized payoff function accepting terminal prices.
}
\description{
Standard terminal payoff functions
}
